Moving permission coordinator presentation into the main coordinator.
