+++ Shared/Coordinator/MainCoordinator+Launch.swift
@@ -8,7 +8,6 @@
 
 import Foundation
 import StreamChat
-import Intents
 
 #if IOS
 extension MainCoordinator {
@@ -31,26 +30,6 @@ extension MainCoordinator {
         self.addChildAndStart(coordinator, finishedHandler: { (_) in })
 
         self.router.setRootModule(coordinator)
-
-        await self.checkForPermissions()
-    }
-
-    @MainActor
-    func checkForPermissions() async {
-        if INFocusStatusCenter.default.authorizationStatus != .authorized {
-            self.presentPermissions()
-        } else if await UserNotificationManager.shared.getNotificationSettings().authorizationStatus != .authorized {
-            self.presentPermissions()
-        }
-    }
-
-    @MainActor
-    private func presentPermissions() {
-        let coordinator = PermissionsCoordinator(router: self.router, deepLink: self.deepLink)
-        self.addChildAndStart(coordinator) { [unowned self] result in
-            self.router.dismiss(source: coordinator.toPresentable(), animated: true)
-        }
-        self.router.present(coordinator, source: self.router.topmostViewController)
     }
 
     func logOutChat() {
